pyproject.toml
[build-system]
requires = [
    "setuptools>=45",
    "wheel",
    "scmver"
]
build-backend = "setuptools.build_meta"

[project]
name = "merlinquantum"
dynamic = ["version"]
description = "Quantum neural network models using photonic circuits - Preview"
readme= "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "MerLin Team"},
    {name = "MerLin", email = "MerLin@quandela.com"}
]

classifiers=[
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Physics",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "torch>=2.0.0",
    "perceval-quandela>=0.13.1",
    "numpy>=2.2.6",
    "pandas",
    "scikit-learn>=1.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-benchmark>=4.0.0",
    "pytest-cov>=5.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
docs = [
    "sphinx>=6.2.1,<7.0.0",
    "sphinx-autodoc-typehints>=1.0.0",
    "sphinxcontrib-bibtex>=2.5.0",
    "renku-sphinx-theme",
    "enum-tools[sphinx]",
    "nbsphinx",
    "nbformat",
    "ipython",
    "ipykernel",
    "gitpython",
    "jinja2==3.0.0",
    "sphinx-multiversion-contrib @ git+https://github.com/Quandela/sphinx-multiversion-contrib.git@1.1.1",
    "scikit-learn>=1.0.0",
]
examples = [
    "matplotlib>=3.5.0",
    "scikit-learn>=1.0.0",
    "jupyter>=1.0.0",
]

[project.urls]
homepage = "https://merlinquantum.ai"
documentation = "https://merlinquantum.ai"
source = "https://github.com/merlinquantum/merlin"
issues = "https://github.com/merlinquantum/merlin/issues"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["merlin", "merlin.core", "merlin.datasets", "merlin.pcvl_pytorch", "merlin.sampling", "merlin.torch_utils"]

[tool.pytest.ini_options]
testpaths = ["tests", "benchmarks"]
python_files = ["test_*.py", "benchmark_*.py"]
addopts = "--cov=merlin --cov-report=term-missing --cov-report=xml --cov-report=html"

[tool.coverage.run]
source = ["merlin"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
    "*/venv/*",
    "*/.venv/*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == \"__main__\":",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
precision = 2
skip_covered = false
# Warning threshold - coverage below this will show warning but won't fail
fail_under = 0
# 80% is our monitoring threshold (configured in workflows)

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"

[tool.ruff]
target-version = "py310"
line-length = 88
preview = true
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    "venv",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "Q",  # flake8-quotes
    "COM", # flake8-commas
    "S",  # flake8-bandit (security)
    "N",  # pep8-naming
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "B904",  # Within an `except` clause, raise exceptions with `raise ... from err`
    "S101",  # Use of assert detected (common in tests)
    "N806",  # Variable in function should be lowercase (allows mixed case)
    "COM812", # Trailing comma missing - handled by ruff format
    "S310",  # Suspicious URL open (urllib.request is safe here)
    "S311",  # Standard pseudo-random generators are fine for non-crypto
    "N803",  # Argument name should be lowercase (allows PhotonicBackend param)
    "N812",  # Imported as non-lowercase (allows import as ML),
    "N801",  # Class name `UnpairedDataset_Quantum` should use CapWords convention
    "N802",  # function name should be lowercase
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["B011"]  # assert False in tests is ok

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

